Trying

+++ myneopixel.py
@@ -76,6 +76,6 @@ brightnesses = [0, 25, 50, 100, 125, 150, 175, 200, 225, 255, 225, 200, 175, 150
 for i in [0, 1, 2, 3]:
     for brightness in brightnesses:
         for pixel in range(101, LED_COUNT):
-            strip.setPixelColor(pixel, Color(0, brightness / 4, 0))
+            strip.setPixelColor(pixel, Color(0, int(brightness / 4), 0))
         strip.show()
         time.sleep(0.02)
